Modify inheritance diagrams so they look better with our api docs.
Remove our copy of inheritance_diagram and use the one that ships with
Sphinx 0.6.
Only show 2 parts of the namespace for objects.  This is an option to the
inheritance-diagram directive, ":parts: 2", which is set in apigen.py.
Set graph and node attributes so diagrams look reasonable for limited screen
space and various graph sizes.



git-svn-id: https://nipy.svn.sourceforge.net/svnroot/nipy/nipype/trunk@479 ead46cd0-7350-4e37-8683-fc4c6f79bf00