Make our tests work with an egg install.  Damn you setuptools!
I lost half my day on this!  Setuptools does a chmod +x on ALL modules
during the install, something distutils does not do.  Nose will not
import executables because it's a security risk, so by default, none
of our tests would run with an egg install.  Need to pass an extra arg
into nose '--exe' to force it to import the executables.  Argh!

List thread on this:
http://www.mail-archive.com/distutils-sig@python.org/msg05009.html




git-svn-id: https://nipy.svn.sourceforge.net/svnroot/nipy/nipype/trunk@530 ead46cd0-7350-4e37-8683-fc4c6f79bf00