ENH:  Added enhanced SEM only interface
This patch allows any program with an SEM interface
to be included in nipype even if it is not part of
Slicer3.  The use of an "launcher" has been
decoupled from the SEM processing, and it is now
possible to use either Slicer3, SlicerQt, BRAINS, or
no launcher at all for generating the nipype
interfaces.