ENH:  In SEM, fileExtensions is a csv list
The fileExtensions attribute is a hint to
the workflow processing system about which file
extensions could be used.  It is intended as
a set of hints for GUI representation, but
the convention is that the first item in the list
is the default behavior.

This list can be thought of like "save as" in Excel
where the first extension would be ".xls", but alternative
options are ".csv", ".txt" etc.