ENH: Whole-graph submission to Condor DAGMan
This is an initial implementation. it should work and should be
reasonable feature-complete. If this gets merged into master it should
replace the qsub-based Condor plugin, as this implementation requires no
additional software (qsub wrapper) it should be more efficient and
versatile.

The code needs to be considered experimental, as it has only been tested
with trivial workflows.

If the old Condor plugin gets removed a documentation update needs to
happen.