ENH: Add opt for converting siemans PAR/REC files
Adds an option for the '-v' flag in dcm2nii, which
switches off the default behavior of converting
all PAR/REC files in a directory. Without this
option, dcm2nii will attempt to convert all files,
some of which may be bad runs and can cause the
entire directory to fail.

This also adds an option in the outfile parser to
search for "-->" which is the syntax returned when
dcm2nii converts PAR/REC files instead of "Saving"
or "Cropping".