ENH: Respect user supplied -o -e flags in qsub_args
The qsub_args may have -o and -e flags pre-specified.
If those are explicitly specified by the user, then do
not use the default values.

In addition, the line was broken apart into units that
make it easier to debug. Initially there was an error
stating that the argument to %d is required to be an
integer.  This failure was removed during the refactoring
process.