ENH: Envirionmental processing for thread limit enforced
The behavior of the number of threads to be used was not
being properly enforced when set outside of nipype.  This
is improper and results in overloading or under utilization
when run in a cluster environment.

This patchset makes the behavior default to the correct
values when run on local or SGE clusters, or if the batch
processing system sets the envirionmental variables.