Add testing for the 'non_daemon' flag of the MultiProc plugin.
The testing scenario spawns 2 subprocesses which spawn each 2 child subprocesses. By default, this fails in Python since subprocesses are not allowed to spawn its' own subproccesses.

We test first if the execution of the scenario fails when the nondaemon_flag is False. This is expected.
Then, we test if the execution succeeds when the nondaemon_flag is True.