Renamed susceptibility pipeline
Since I'm to replace this workflow, I rename it with its ``classic''
name. The new pipeline will be based on the latest versions of
```fsl_prepare``` and ```epi_reg``` of FSL.

Additionally, my intention is to provide with an interface for
the new tool ```topup``` and a new workflow for T2-to-B0 registration-based
correction.