BUG: antsRegistration command line failure
The antsRegistration program does not have a
named sampling strategy called "Dense".  In
order to get the 1 sample per voxel sampling
strategy (aka dense sampling) the user must
either not supply anything, or explicitly
supply the "None" name.

Perhaps this should have been called the
"subSamplingStrategy", but that is too
invasive of a change to ants.