ENH: Adding new ants command line feature.
A newer version of ANTs has a command line feature
added that allows for more optimized processing in
multi-stage registrations.

ENH: Added new state saving interface to ANTs

In order to improve registration efficiency, it is important that
we can save and restore the state of a registration so taht
continuation may occur from a previous state.

ENH: Make transform writing mutually exclusive

Writing composite or non-composite transforms needs
to be mutually exclusive in the new version of
ants.