ENH: A set of autogenerated interfaces
This is a set of automatically generated interfaces
to provide the nodes necessary to for running the
BRAINSAutoWorkup pipeline.  The main source code
for the binaries related to this interface is
located at: https://github.com/BRAINSia/NAMICExternalProjects

This is a parallel sytem to the Slicer GUI that provides many
the consistently performing tools in a command line project
without the need for complicated extension building and shared
library dependancy manipulation.

This set of interfaces was automatically generated with
the following psuedocode set of instructions:

cd ~/
git clone https://github.com/BRAINSia/NAMICExternalProjects.git
mkdir ~/NEP-bld
cd ~/NEP-bld
ccmake ~/NAMICExternalProjects.git
make

~/NEP-bld/BRAINSTools/AutoWorkup/local_generate_classes.py