ENH: Attempt to use hard links for data sink.
In many cases on Unix, the data sink is on the same
physical drive as the internal nipype cache.  In
that case, we can use a hard link to save both
time necessary to duplicate data, and space
necesssary to hold the same data at two different
inodes.  This allows removal of the cache
directory without modifying the results
directory.

In large analysis, this optimization can save
several terabytes of storage consumption.