ENH: Remove temporary test files.
As new patches are created, testing should be done
before a PR is requested.

The recommended testing is
make specs; python setup.py install; make check-before-commit

In this instance, many temporary test files are created
that obscure the view of the code that was actually changed.

The temporary test files (which may change during test runs)
should NOT be included in the repository.

Step 1: Remove temporary files
  find . -name "test_auto*.py" -exec git rm {} \;
  git rm nipype/testing/data/von_errmap.nii.gz -f
  git rm nipype/testing/data/von-ray_errmap.nii.gz -f

Step 2: Add file patterns to .gitignore