Makes S3DataGrabber change all paths from s3 to local
Previously, S3DataGrabber checked if outputs were lists or strings,
and changed them in those cases. However, if output was a unicode,
then it went unchanged. That's clearly not what we want, so we now
change in all cases: iterating through all values of output if it has
__iter__ attribute (is list-like), and using the value as the path if
it doesn't.