ENH: Replicates N4BiasFieldCorrection output style
This replicates the method from in N4BiasFieldCorrection used to
specify more than one output file (as suggested by blakedewey).
It updates the doctest and test_auto_DenoiseImage.py with new
variables save_noise and noise_image. It alse removes the version,
help, and short_help options since nipype is a pipelining tool
and a pipeline would not need to see the output if those options
are set.