Update brainsuite.py
I updated the formatting to comply with the PEP8 style guide (The exception is that some lines run over 79 characters)

As for instantiating names, the reason I approached it this way is to shorten the names of outputs that ran through multiple steps of processing using brainsuite command line tools. Is it fine to leave the instantiation of outputs how it is now?