FIX: Fixes the inputs of SmoothTesselation to be consistent with the documentation for 'mris_smooth'
The only positional arguments according to the documentation should be in_file and out_file.
The position for these inputs should be -2 and -1, respectively.
Other options do not require a fixed position.
Defaults are set automatically by 'mris_smooth' and do not need to be written into the nipype wrappings.