[ENH] Add a FramewiseDisplacement interface
Calculates the FD following the implementation of fsl_motion_outliers,
that theoretically follows that of Power et al. 2012.

If ```save_plot``` is ```True```, then it will save an FD plot using
seaborn (but this is optional).

Includes a test that assess that this implementation is equivalent
to the one in fsl_motion_outliers