Fixing 0 value given to executable
Hi,

I realized that two operations don't work if you use 0 as a value (thr and uthr). The interface set the value to 0.000000 when you give zero as an operand value. The executable failed if you give zero with decimals. 

I added some extra lines of code in _format_arg to check wheter or not the value is 0. If so, return the value zero without decimal.

Kind Regards,

Ben