Cast DVARS float outputs to avoid memmap error
When the input files are not compressed, numpy wraps memmap around the
dtype. This makes the traits.Float validation to fail.

An more robust way would be overloading traits.Float to validate also
these memmap objects. But this is probably overkilling the problem.

For now, this casting is the easiest and safest solution.