[ENH] Revise how docker build is run in CircleCI
Adds the ```--rm=false``` arguments so that docker does not remove
intermediate layers after build. This should speed up the build process
and allow Docker to export intermediate layers into the cache.

Also, the ~/docker folder has been added as an artifact to check if
the cache tarfile is generated.