[FIX] CircleCI tests - error codes are not propagated
The former structure of the tests.sh file in which all the tests are called
did not exit with error if one of the test commands errored, since the
last exit code was from codecov, and no exit code tracking or propagation
was done.

This PR fixes that issue. Additionally, the scripts related to CircleCI
have been split up into a new .circle/ folder.