Use copyfile to link the surface context
Mimics the FreeSurfer MRIsBuildFileName to select appropriate files
If copying files, defines pial and thickness_name (as needed) to ensure
copy.

Adds a non-argument sphere trait that MUST NOT be changed. mris_expand
will not run in a node if there is not a `?h.sphere` associated with the
in_file.