Minor QoL change for write_graphs()
changed behavior of write_graph() slightly. If workflow name is defined but not base_dir the dot graphs be stored in a directory named after the workflow name, rather than just dumping to the current working directory. If neither name nor base_dir is defined then it should default to current behavior.