FIX: Use un-copied subject directory in MRIsExpand
mris_expand expects the entire subject directory to be available. This
change abuses _get_filecopy_info in a different way from before,
changing the output filename only in cases where mris_expand is to be
run in a node.

Fixes issue in which absent ../mri/transforms/talairach.* prevents xform
from being written to the new surface.