fix: use gzip compression -6 instead of -1
The lowest amount of compression created a tar.gz file of over 5.0 GB. Persisting this to the CircleCI workspace did not show errors, however, tar.gz file in the next step of the workflow would be empty. This caused `docker load` to fail. More compression leads to a tar.gz file that is about 4.7 GB. This relatively smaller file can be loaded in the next step of the CircleCI workflow.

In a future PR, the Nipype container will be minimized with ReproZip. This will create smaller containers and alleviate the issue that this commit aims to fix.