ENH: Parse Boutiques output structure
Adds support for dynamic parsing of boutiques output-files spec.

Unfortunately, boutiques does not support specification of which input
options will lead to the generation of specific output files, so _all_
output files are being specified. This will likely lead to some problems
down the road, and might be worth suggesting to the boutiques team as a
future enhancement of their schema.