[RF] Detachable ``run_command`` process
**Introduction**: the current implementation of ``run_command`` makes
use of ``Stream``s to redirect output in a sort of inflexible way, and
does not allow running the processes in the background.

**This PR**: modifies the behavior of ``run_command`` opening a new
self-control thread, that is in charge of printing to the terminal if
the ``terminal_output`` is set to ``'stream'``. Output streams are
redirected to files by default, trying to reduce the amount of data
buffered in memory and also to ensure that there is a trace of nipype's
execution. Additionally, a new argument ``background`` can be set to
``True``, detaching from the process. This lends itself to important
improvements on the ``MultiProc`` variants, as all these processes can
be launched directly from the main thread and run in the background.
I hope this will drastically reduce the memory fingerprint of workers
(if we actually need to maintain a ``Pool``, which is probably not the
case anymore!).

**Tests**: I've added a couple of tests, but I'm happy to add more as I
am aware that this may be too big of a change to the nipype interfaces.

**Extra**: removed a couple of deprecated objects that we promised to
trash past nipype-1.1.0.