fix(Python<3.5): new test ``test_modify_paths_bug`` raising PicklingError
When ``input_spec`` and/or ``output_spec`` are defined as private
classes, pickle has a hard time to find the right type and breaks with
Python 2 and 3.4.

This problem of pickle has been exposed by the new way of pickling
InterfaceResults.

I believe we can live with such a restriction - i.e.,
(In/Out)put specs must not be scoped
within the interface (and again, only for Python<3.5).

cc/ @satra