ENH: Minimize the number of calls to ``_load_results`` when populating inputs
This PR attempts to alleviate #3014 by opening the result file of a
source node only once when that node feeds into several inputs of the
node collecting inputs.
Before these changes, a call to ``_load_results`` was issued for every
input field that needed to collect its inputs from a past node.
Now, all the inputs comming from the same node are put together and the
``_load_results`` function is called just once.

The PR also modifies the manner the ``AttributeError``s (#3014) were handled
to make it easier to spot whether an error occured while loading results
araises when gathering the inputs of a node-to-be-run or elsewhere.