fix: improve error handling of ``CommandLine`` interfaces
Currently, errors arising from interpolating the command line of these
interfaces is handled poorly at the ``Node`` level.
If the command line cannot be built, the error is printed in the logfile
but the exception is caught and never raised (i.e., likely leading to an
infinite loop as the execution is not stopped).

I have experienced that while debugging fMRIPrep.
To learn which of the inputs of a faulty interface derived from
``CommandLine`` was not being formatted, I had to also add the error
annotation proposed for the ``_parse_inputs`` inner loop.