ENH: Explicitly specify write_text encoding format
When writing reports from nipype the write_text
function could default to using 'ascii' coding
for writing the report.  This would fail if any
part of the report tried to generate a unicode
character.

The issue was encountered when the OS environmental
variable "PS1" had a non-ascii character in the
environmental variable.

The result was a failure of the nipype pipeline
from running due to an exception being thrown
while trying to write the debugging information.