FIX: Normalize constant column to 0th
For Legendre regressors, the ith column is the ith-order polynomial, so
the constant column is 0. For the cosine regressors, a constant column
was appended to the end, in contradiction of the docstring. This brings
both into alignment so columns are sorted from lowest to highest
frequency and aligns the DCT behavior with its docstring.